Christoph Niedermaier <cniedermaier@dh-electronics.com> says:

Add the "-l" parameter to the "env select" command to print the available
environment targets, convert all argument parsing in the file nvedit.c to
getopt() and check that with unit tests.

The first patch converts the parsing of arguments to getopt() for all env
commands. The second one adds unit tests for checking the env command.
These were used to check the env commands before and after the conversion.
The third patch adds the actual new parameter "-l" for the "env select"
command.

Link: https://lore.kernel.org/r/20250630151232.39100-1-cniedermaier@dh-electronics.com

+++ drivers/clk/clk-cdce9xx.c
@@ -103,7 +103,7 @@ static int cdce9xx_clk_probe(struct udevice *dev)
 	u32 val;
 	struct clk clk;
 
-	val = (u32)dev_read_addr_ptr(dev);
+	val = (ulong)dev_read_addr_ptr(dev);
 
 	ret = i2c_get_chip(dev->parent, val, 1, &data->i2c);
 	if (ret) {
@@ -226,10 +226,10 @@ static ulong cdce9xx_clk_set_rate(struct clk *clk, ulong rate)
 }
 
 static const struct udevice_id cdce9xx_clk_of_match[] = {
-	{ .compatible = "ti,cdce913", .data = (u32)&cdce913_chip_info },
-	{ .compatible = "ti,cdce925", .data = (u32)&cdce925_chip_info },
-	{ .compatible = "ti,cdce937", .data = (u32)&cdce937_chip_info },
-	{ .compatible = "ti,cdce949", .data = (u32)&cdce949_chip_info },
+	{ .compatible = "ti,cdce913", .data = (ulong)&cdce913_chip_info },
+	{ .compatible = "ti,cdce925", .data = (ulong)&cdce925_chip_info },
+	{ .compatible = "ti,cdce937", .data = (ulong)&cdce937_chip_info },
+	{ .compatible = "ti,cdce949", .data = (ulong)&cdce949_chip_info },
 	{ /* sentinel */ },
 };
